+++ rpmGauge.py
@@ -34,10 +34,10 @@ class Rpm(QWidget):
         self.rpmLabel.move(320,120)
         self.rpmLabel.show()
 
-    @pyqtSlot(int)
+    @pyqtSlot(float)
     def rpm_update(self, value):
         #self.rpmValue = value
-        self.rpmValue = (800 * self.rpmValue + (1024 - 800) * value) >> 10
+        self.rpmValue = (800 * int(self.rpmValue) + (1024 - 800) * value) >> 10
         if self.rpmValue < 0:
             self.rpmValue = 0
         self.rpmLCD.display(str(10*int(self.rpmValue/10)).zfill(4))#displays rpm, reduces precision to imporve readablity
